guarantee ranges
- strtoll for guaranteed u32/s32-inclusive range
- changed various types to s64 for the same reason
- removed errno
  - the max buffer size of `S32_MAX_SIZE` ensures that a 64 bit overflowing value can never be reached

+++ src/pc/djui/djui_input_number.h
@@ -16,7 +16,7 @@ struct DjuiInputNumber {
     enum InputNumberType type;
     u8 digits;
     s32 min, max;
-    void *value; s32 saved;
+    void *value; s64 saved;
     bool valid;
 };
 
@@ -29,4 +29,4 @@ void djui_input_number_text_change(struct DjuiBase *caller);
         s32 *: NUMTYPE_S32 \
     ), min, max)
 
-struct DjuiInputNumber *_djui_input_number_create(struct DjuiBase *parent, void *value, enum InputNumberType type, s32 min, s32 max);
+struct DjuiInputNumber *_djui_input_number_create(struct DjuiBase *parent, void *value, enum InputNumberType type, s64 min, s64 max);
