+++ src/pc/djui/djui_inputbox.c
@@ -62,8 +62,8 @@ void djui_inputbox_set_text(struct DjuiInputbox* inputbox, char* text) {
     snprintf(inputbox->buffer, inputbox->bufferSize, "%s", text);
 }
 
-void djui_inputbox_set_number(struct DjuiInputbox* inputbox, s32 value) {
-    snprintf(inputbox->buffer, inputbox->bufferSize, "%d", value);
+void djui_inputbox_set_number(struct DjuiInputbox* inputbox, s64 value) {
+    snprintf(inputbox->buffer, inputbox->bufferSize, "%lld", value);
 }
 
 void djui_inputbox_select_all(struct DjuiInputbox* inputbox) {
